requirements.txt
python-dotenv
groq
pillow
opencv-python
pyperclip
google-generativeai